Model changed to exclude dependency on androidx. Files renamed to reflect purpose. Packages renamed and restructured to support future extensions.
